validating screenshotAttachment before changing name (#125)
* adjust exui submit event method to reduce flakiness

* fixed lint issues

* fix attempt retry submit

* add prettier, update eslint, add one class page page architecture, add husky precommit, add full functional tests for each browser, improve accessibility testing

* new commit

* checking husky

* checking husky

* testing husky

* testing husky

* configuring attachments in accessibility tests

* lint changes

* add custom eslint plugin

* yarn install new custom eslint plugin

* fixed errors check issue

* fix

* fixes to pipeline behaviour

* reverting jenkins nightly

* fix

* fixed retry click submit

* increasing retry timeout for expect

* changing .catch() to try/catch

* increasing timeouts

* update test configuration

* fixed Jenkinsfile_nightly

* fixed jenkins nightly

* Created expectNoText and expectNoSelector Methods in BasePage

* more test configuration set up

* Add more details to tests

* last bit of configuration

* prettier changes

* final configurations

* update

* prefer-step-decorator lint, AxeCacheHelper

* last lint fixes

* updated yarn lock

* update jenkins

* final configurations

* changing screenshot name to failed.png

* check if status is failed

* validating screenshot attachment and video

+++ playwright/playwright-fixtures/utils/test-utils-fixtures.ts
@@ -24,11 +24,11 @@ export const test = base.extend<TestDataFixture>({
       testInfo.errors.length > 0
         ? testInfo.errors.every((error) => error.value === 'accessibility')
         : false;
-    if (allErrorsAxe) {
+    if (allErrorsAxe && screenshotAttachment && page.video()) {
       FileSystemHelper.delete(screenshotAttachment.path, { force: true });
       FileSystemHelper.delete(await page.video().path(), { force: true });
       test.fail();
-    } else {
+    } else if (screenshotAttachment) {
       await testInfo.attach('failed.png', { path: screenshotAttachment.path });
       FileSystemHelper.delete(screenshotAttachment.path, { force: true });
     }
